Add javascript, typescript and react support
In Emacs 27 js2-mode has support for JSX, so its enabled as a minormode
when web-mode or js-mode mayor modes are activated.

Tide provides completion, linting and formatting for Javascript,
Typescript and JSX. It needs tide server as a dependency and a
jsconfig.json file per project. Check the repository for more info.